Dodana kontrola prazne košarice i cijena je promijenjena iz tipa int u float. #6

+++ FOOD2GO/webservice/src/main/java/hr/foi/air/webservice/WebServiceCaller.java
@@ -127,7 +127,7 @@ public class WebServiceCaller {
         CallFromServer("dodajStavkeNaRacun");
 
     }
-    public void DodajCijenuRacunu(Racun racun, int cijena) {
+    public void DodajCijenuRacunu(Racun racun, float cijena) {
         WebService webService = retrofit.create(WebService.class);
         call=webService.DodajCijenuNaRacun(racun.getID(),cijena);
         Log.i("AIR_WebServiceCaller",String.valueOf(cijena));
